// payload-types.ts
export interface Page {
  id: string;
  title: string;
  slug: string;
  layout: (
    | {
        heading: string;
        subheading: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        image: string | Media;
        cta_button: {
          label: string;
          url: string;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'hero';
      }
    | {
        heading: string;
        subheading?: string | null;
        image: string | Media;
        content: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        cta: {
          label: string;
          url: string;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'content';
      }
    | {
        heading?: string | null;
        form: string | Form;
        id?: string | null;
        blockName?: string | null;
        blockType: 'newform';
      }
  )[];
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}